Refactor accelerate test suite, add amplify tests
