+++ recipes/esm2_accelerate/hydra_config/L0_sanity_amplify.yaml
@@ -2,11 +2,11 @@ defaults:
   - defaults
   - _self_
 
-model_tag: "nvidia/esm2_t6_8M_UR50D"
+model_tag: "nvidia/AMPLIFY_120M"
 stop_after_n_steps: 250
 
 trainer:
-  run_name: "esm2_t6_8M_UR50D_sanity"
+  run_name: "amplify_120M_sanity"
   per_device_train_batch_size: 2
   per_device_eval_batch_size: 2
   save_steps: 1000
@@ -14,4 +14,3 @@ trainer:
   logging_steps: 10
   report_to: "none"
   dataloader_num_workers: 4
-  warmup_steps: 0
